Squashed commit

+++ src/main/java/com/ase/restservice/service/AccountService.java
@@ -4,15 +4,19 @@ import com.ase.restservice.exception.AccountAlreadyExistsException;
 import com.ase.restservice.exception.AccountNotFoundException;
 import com.ase.restservice.model.Account;
 import com.ase.restservice.repository.AccountRepository;
+
 import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import static com.ase.restservice.ApplicationSecurity.getUsernameOfClientLogged;
 
 /**
  * Service for Account operations.
  */
 @Service
-public class AccountService implements AccountServiceI {
+public class AccountService implements com.ase.restservice.service.AccountServiceI {
 
   @Autowired
   private AccountRepository accountRepository;
@@ -31,10 +35,13 @@ public class AccountService implements AccountServiceI {
 
       // reach here means the account already exists in database
       throw new AccountAlreadyExistsException(
-        "Account already exists with ID :: " + accountId
+          "Account already exists with ID :: " + accountId
       );
     } catch (AccountNotFoundException e) {
       // reach here means account does not already exist in database
+      //clientID placement must be handled by system.
+      String clientID = getUsernameOfClientLogged();
+      account.setClientId(clientID);
       return accountRepository.save(account);
     }
   }
@@ -46,6 +53,7 @@ public class AccountService implements AccountServiceI {
    * @return Updated account
    */
   public Account updateAccount(Account account) throws AccountNotFoundException {
+    //TODO might cause an security risk as users shouldnt be able to change the client
     try {
       String accountId = account.getAccountId();
       Account dbAccount = this.getAccountById(accountId);
@@ -63,7 +71,13 @@ public class AccountService implements AccountServiceI {
   public void deleteAccountById(String accountId) throws AccountNotFoundException {
     try {
       Account dbAccount = this.getAccountById(accountId);
-      accountRepository.deleteById(accountId);
+      //TODO check if i can delete it
+
+      String clientID = getUsernameOfClientLogged();
+
+      if (clientID.equals(dbAccount.getClientId())) {
+        accountRepository.deleteById(accountId);
+      }
     } catch (AccountNotFoundException e) {
       throw new AccountNotFoundException(e);
     }
@@ -84,7 +98,9 @@ public class AccountService implements AccountServiceI {
   }
 
   /**
-   * List all accounts.
+   * THIS SHOULD NOT BE USED DUE TO SECURITY.
+   * THIS HAS TO BE HERE FOR COMPILER
+   * Lists all accounts.
    *
    * @return list of accounts
    */
@@ -96,8 +112,8 @@ public class AccountService implements AccountServiceI {
    * Method to increase or decrease a user's account balance.
    *
    * @param accountId Primary key of account
-   * @param amount dollar amount to change account balance by. If negative, will decrease the
-   *               account balance.
+   * @param amount    dollar amount to change account balance by. If negative, will decrease the
+   *                  account balance.
    * @return Returns the account with the updated balance
    * @throws AccountNotFoundException if account does not exist in the database
    */
@@ -107,6 +123,14 @@ public class AccountService implements AccountServiceI {
     account.setBalance(account.getBalance() + amount);
     final Account updatedAccount = this.updateAccount(account);
     return updatedAccount;
+  }
+  /**
+   * List all accounts that logged in client owns.
+   *
+   * @return list of accounts
+   */
+  public List<Account> listAllAccountsOfClient(String clientId) {
+    return accountRepository.findAllAccountByClient(clientId);
   }
 
 }
